fix portability concerns with size_t printfs
Apparently size_t is a concern for printf, as %lu may not be the correct size
to printf as on all platforms. There are two options to address this:
 - cast size_t to unsigned long
 - use %zu

This commit uses the former option, as the second option is apparently also
nonstandard for compilers, and also is dependent on C99. Casting seems to be
the "most" portable option, other than not using size_t, printf, or probably C.

+++ src/authentication_2.c
@@ -39,7 +39,7 @@ unpack_authenticated_variable (const auth_data_t *auth_data, timestamp_t *timest
   if (auth_data->auth_msg_size < sizeof (auth_info_t))
     {
       prlog (PR_ERR, "Buffer too small for an auth2 header - got %lu bytes.\n",
-             auth_data->auth_msg_size);
+             (unsigned long) auth_data->auth_msg_size);
       return SV_BUF_INSUFFICIENT_DATA;
     }
 
@@ -67,7 +67,8 @@ unpack_authenticated_variable (const auth_data_t *auth_data, timestamp_t *timest
    */
   if (da_length <= sizeof (auth_cert_t))
     {
-      prlog (PR_ERR, "da_length in auth header too short for fixed data - %lu bytes\n", da_length);
+      prlog (PR_ERR, "da_length in auth header too short for fixed data - %lu bytes\n",
+             (unsigned long) da_length);
       return SV_AUTH_SIZE_INVALID;
     }
